Adiciona materialize no projto e estiliza telas
