+++ docker/docker-compose.yml
@@ -2,7 +2,7 @@ version: "3"
 services:
   frontend:
     tty: true
-    container_name: application_front_end
+    container_name: application_reactjs_alura
     build: ./
     ports:
       - "3000:3000"
